:bug: fix: 베팅 시, Point 차감되도록 수정, 진행중인 배틀에만 베팅할 수 있도록 수정

+++ backend/ninucco/src/main/java/co/ninuc/ninucco/db/entity/Member.java
@@ -73,7 +73,7 @@ public class Member implements UserDetails{
     public void updateElo(int elo){
         this.elo =elo;
     }
-    public void updatePoint(int point){
+    public void updatePoint(long point){
         this.point =point;
     }
